fix: surface SECURE+ thor address and decimal amounts on verify (#4333)
Teach the memo parser to recognise `SECURE+:<thor address>` so the
joined-device verify screen for Mint Secured Asset surfaces the user's
thor address and a Mint operation row, matching the initiator and the
extension. Also switch the Amount row gate from BigInteger to
BigDecimal so sub-1 amounts like 0.0007 render instead of being
silently hidden by the NumberFormatException catch.

+++ app/src/main/java/com/vultisig/wallet/ui/screens/deposit/VerifyDepositScreen.kt
@@ -53,7 +53,7 @@ import com.vultisig.wallet.ui.screens.swap.VerifyCardDetails
 import com.vultisig.wallet.ui.screens.swap.VerifyCardDivider
 import com.vultisig.wallet.ui.theme.Theme
 import com.vultisig.wallet.ui.utils.asString
-import java.math.BigInteger
+import java.math.BigDecimal
 
 @Composable
 internal fun VerifyDepositScreen(
@@ -214,7 +214,7 @@ internal fun VerifyDepositScreen(
                     if (
                         tx.token.value.isNotEmpty() &&
                             try {
-                                tx.token.value.toBigInteger() > BigInteger.ZERO
+                                tx.token.value.toBigDecimal() > BigDecimal.ZERO
                             } catch (e: Exception) {
                                 false
                             }
